Remove v2 of the DAML model and add v3 instead (#35)
I have deliberately not put the project root of the model at the root of the
repo. I have also deliberately checked in the DAR. The rationale behind this
is that v4 of the model will need access to v3 of the model for generating
the upgrading models I'm fully aware that this is against the spirit of VCSs,
but we simply do not yet have the tools to use the kind of development
workflow we're used from other languages. The main reasom for this project
to exist is that we need to figure out how to actually do that. I'd like to
avoid shooting ourselves in the foot in that endeavour on day one by beeing
to dogmatic about how development workflows should look like. Let's clean
up the mess we make on day 3 after developing the right tools on day 2.