add checked-in dars to sandbox image
Note: the sandbox image still depends on `v3/daml.yaml` for now because
that is what specifies the DAML version. If we're heading towards a
model where we might have multiple daml "projects" inside the same repo
we may need to think about another, separate way to specify the DAML
version we use for production services (this applies to json-api as
well).