bind sandbox and jso-api to 0.0.0.0; deploy dars
0.13.42 changed the default from 0.0.0.0 to localhost for sandbox.
json-api is still defaulting to 0.0.0.0 for now, but by making it
explicit we make our deployment future-proof in case it follows in the
steps of the sandbox.

This also changes the deployment process to only try to upload *.dar
files to the sandbox, to avoid trying to upload the README.md file that
has been added to the deployed folder.