Add boilerplate for an upgrade package (#132)
This PR adds a new v4 directory which is identical to v3 except for
the SDK version which is set to 0.13.46. The resulting DAR is added at
`released/davl-v4.dar`.

In addition to that, it adds a `v4-upgrade` directory which also
compiles against 0.13.46 and uses v3 and v4 via data-dependencies. I
included the boilerplate to actually make this compile and to be able
to create v4 contracts and archive v3 contracts.

The actual upgrade contracts are not yet included, I’ll leave that to
someone else or at least for a separate PR.

I’ve also added signatory functions which are hand-written.