Upgrade DAML SDK used for Platform and UI to 0.13.51
We're using `@daml/types` and `@daml/ledger` from npmjs.org now.

I've deliberately not switched to `@daml/react` from nmpjs.org because I
want a playground to easily experiment with new features for this libary.
The files in `ui/src/daml-react-hooks/` are up-to-date with the ones of
`@daml/react` on npmjs.org though.