Add and use a React hook calling the streaming query JSON API (#155)
This also allows us to entirely rewrite the existing `useQuery` and
`useFetchByKey`. There's no more need to have the Redux reducer and one
central store since users who want updates can use `useStreamQuery` instead.
This removes a lot of complexity. Now, every instance of the hooks has
its own, very simple data store implemented using `useState`.

Unfortunately, `useStreamFetchByKey` is not yet possible since the
correspoding endpoint in the JSON API. Thus, to see the updates to a
contract key an explicit reload is still necessary.

All of this has still plenty of quirks when it comes to exceptions and
asynchronicity. However, nothing of this was better before. I'll address
it in a bunch of follow-up PRs.