Upgrade to DAML SDK 0.13.53
Since `daml2ts` in the new SDK supports multiple DARs in one call, I've
decided to step checking the generated TS in but rather add an instruction
to the README on how to (re)generate it.

I've also not switched to upstream `@daml/react` although we now can finally
do this. I'll do so in a follow-up PR.