Garbage collect DAVL v3 and the upgrade from v3 to v4 (#231)
Nothing in the repo references v3 or the upgrade from v3 to v4 anymore.
There are also no more active contracts frpm v3 and the only ones still
active from the upgrade might be the upgrade agreements. They are quite
pointless without any active data from v3 but archiving them isn't as
easy as it sounds since you need the approval from the employee and the
company to do so. That is something we should rethink when designing
the next upgrade.

Anyway, there's no point in keeping the v3 models or the upgrade
models here. git has all the history in case we need it.