db backups
This PR aims at adding backups to the DAVL project. We've been running
without a safety net for long enough now.

Because we already have a database in production, and because Terraform
tends to replace machines by deleting the existing ones, a bit of care
will be needed in deployment. The plan is to:

0. Create a bucket to store backups.
1. Manually create a backup from the existing DB, upload to bucket.
2. Create a new machine, backed-up-db, that behaves as the current DB
   but:
  1. Downloads and restores the latest snapshot on startup.
  2. Creates and uploads a new snapshot every hour.
3. Wait a bit to check everything seems fine with the new machine; in
   particular, check that:
  1. It does indeed upload a new snapshot every hour.
  2. It does indeed restore from the latest server. This might involve
     manually killing the machine.
4. Shut down the DAVL website (ui and ledger machines).
5. Manually create a backup from the existing DB, upload to bucket.
6. Delete existing machine, reset backed-up-db (which at this point will
   load from the manual backup).
7. Restart ledger and ui machines, but having the ledger machine connect
   to the backed-up-db machine.

At the time of opening, the PR contains a single, untested draft of
steps 0 and 2.