fix database backup names (#251)
There is a typo in the current date format, which means backups will not
sort correctly so we may end up loading the wrong version on restart.

Deploying this will be a little bit tricky. Steps needed:

1. Manually shut down the website (i.e. kill the ledger and ui machines)
2. Manually SSH into the current db machine, change the date format, and
   trigger a backup.
3. Ensure that manual backup is the latest one as per sort order. This
   may involve renaming or deleting existing backups.
4. Run `terraform apply`, which should kill the db machine and then
   recreate it with the correct date format, and load the manual backup.

As there's no reason to keep it around anymore, this PR also removes the
old db, which will be deleted by step 4 above.