add dev-env
I got mildly annoyed at having to #259 again, and then @cocreature
pushed me over the edge, so here is a proper fix.

Note: this only include Terraform in the dev-env for now, but we should
probably look at expanding that over time.

This also means Terraform gets bumped again, to 0.12.28, because I was
too lazy to hunt down the latest nixpkgs snapshot that still included
0.12.26, so instead I just picked the latest one.

Also note that nix-bundled Terraform comes with its embedded plugins,
which do not have a version string, so we need to remove the version
requirement from the tf file.