restart Docker containers after Docker failure
I'm not super happy about this. The issue is that our Ubuntu machines
come with an automated daily `apt-get upgrade`, which on the face of it
seems like a good idea. However, on a Docker upgrade, the process seems
to be to SIGTERM all the containers, then shut down the daemon, install
the upgrade, and restar the daemon. If you think there's a missing step
there, so do I.

There are two bad solutions:

- Disable automated upgrades. I think that's going to be hard sell with
  the security team, but maybe we don't need to tell them. (hi
  @nycnewman :wave:).
- Restart containers on failure, which is what this PR does. The
  annoying part about it is that it may mask "real" failures, which sort
  of defeats the point of having a test environment like DAVL.

I'm not entirely sure this will work: the containers do get a non-0 exit
code, but it's unclear to me exactly how Docker handles its own restart
and whether this will be enough for it to restart our containers. It may
be we need to change `on-failure` to `always`. The [documentation] is a
bit unclear about that.

[documentation]: https://docs.docker.com/engine/reference/run/#restart-policies---restart